[dashboard builder] move component/*Container => containers/*
