+++ superset/assets/src/dashboard/stylesheets/components/chart.less
@@ -31,13 +31,13 @@
   .resizable-container:hover
   > .dashboard-component-chart-holder:after,
 .dashboard--editing .dashboard-component-chart-holder:hover:after {
-  box-shadow: inset 0 0 0 1.4px @gray-light;
+  border: 1px solid @gray-light;
 }
 
 .dashboard--editing
   .resizable-container.resizable-container--resizing:hover
   > .dashboard-component-chart-holder:after {
-  box-shadow: inset 0 0 0 1.4px @indicator-color;
+  border: 1px solid @indicator-color;
 }
 
 .dashboard--editing
@@ -52,7 +52,7 @@
   .dashboard-component-chart-holder:hover
   .dashboard-chart
   .chart-container {
-  opacity: 1;
+  opacity: 0.7;
 }
 
 .dashboard--editing
