+++ superset/assets/src/dashboard/stylesheets/components/column.less
@@ -25,7 +25,7 @@
   > .grid-column:after,
 .dashboard--editing .grid-column:hover:after {
   border: 1px dashed @gray-light;
-  box-shadow: 2px 2px 5px 1px @gray-light;
+  box-shadow: 0 4px 4px 0 rgba(0, 0, 0, 0.1);
 }
 
 .grid-column > .hover-menu--top {
