+++ superset/assets/src/dashboard/stylesheets/components/row.less
@@ -31,7 +31,7 @@
   > .grid-row:after,
 .dashboard--editing .grid-row:hover:after {
   border: 1px dashed @gray-light;
-  box-shadow: 2px 2px 5px 1px @gray-light;
+  box-shadow: 0 4px 4px 0 rgba(0, 0, 0, 0.1);
 }
 
 .grid-row.grid-row--empty {
