+++ superset/assets/src/dashboard/stylesheets/dashboard.less
@@ -1,4 +1,4 @@
-@import "./less/cosmo/variables.less";
+// @import './less/cosmo/variables.less';
 
 .dashboard .chart-header {
   position: relative;
@@ -10,7 +10,7 @@
 
   .dropdown-menu.dropdown-menu-right {
     right: 7px;
-    top: -3px
+    top: -3px;
   }
 }
 
@@ -26,7 +26,8 @@
     color: red;
   }
 
-  &:hover, &:focus {
+  &:hover,
+  &:focus {
     background: none;
     cursor: pointer;
   }
@@ -57,7 +58,8 @@
   overflow-y: auto;
 }
 
-.hidden, #pageDropDown {
+.hidden,
+#pageDropDown {
   display: none;
 }
 
@@ -75,21 +77,22 @@
 
 .dashboard .title .favstar {
   font-size: 20px;
+  line-height: 1em;
   position: relative;
   top: -5px;
 }
 
 .ace_gutter {
-    z-index: 0;
+  z-index: 0;
 }
 .ace_content {
-    z-index: 0;
+  z-index: 0;
 }
 .ace_scrollbar {
-    z-index: 0;
+  z-index: 0;
 }
 .slice_container .alert {
-    margin: 10px;
+  margin: 10px;
 }
 
 i.danger {
